requirements.txt
# Required for tripper_app_like_nav.py and dash_ui:
cryptography>=41.0          # RSA/AES auth handshake with the dash

# Required only for the dash_ui prototype (interactive UI):
pygame-ce>=2.5              # community pygame fork; works on Python 3.13/3.14
